refactor(course-info): server-side rendering
