+++ package.json
@@ -30,6 +30,7 @@
     "react-dom": "19.1.0",
     "react-resizable-panels": "^3.0.4",
     "recharts": "2.15.4",
+    "swr": "^2.3.6",
     "tailwind-merge": "^3.3.1"
   },
   "devDependencies": {
